+++ src/server/game/AuctionHouseBot/AuctionHouseBotBuyer.cpp
@@ -101,7 +101,7 @@ uint32 AuctionBotBuyer::GetItemInformation(BuyerConfiguration& config)
     time_t now = GameTime::GetGameTime();
     uint32 count = 0;
 
-    AuctionHouseObject* house = sAuctionMgr->GetAuctionsMap(config.GetHouseType());
+    AuctionHouseObject* house = sAuctionMgr->GetAuctionHouseByFactionTemplateId(config.GetHouseType());
     for (AuctionHouseObject::AuctionEntryMap::const_iterator itr = house->GetAuctionsBegin(); itr != house->GetAuctionsEnd(); ++itr)
     {
         AuctionEntry* entry = itr->second;
@@ -255,7 +255,7 @@ void AuctionBotBuyer::PrepareListOfEntry(BuyerConfiguration& config)
 void AuctionBotBuyer::BuyAndBidItems(BuyerConfiguration& config)
 {
     time_t now = GameTime::GetGameTime();
-    AuctionHouseObject* auctionHouse = sAuctionMgr->GetAuctionsMap(config.GetHouseType());
+    AuctionHouseObject* auctionHouse = sAuctionMgr->GetAuctionHouseByFactionTemplateId(config.GetHouseType());
     CheckEntryMap& items = config.EligibleItems;
 
     // Max amount of items to buy or bid
@@ -411,7 +411,7 @@ void AuctionBotBuyer::BuyEntry(AuctionEntry* auction, AuctionHouseObject* auctio
 
     // Remove auction item and auction from memory
     sAuctionMgr->RemoveAItem(auction->itemGUIDLow);
-    auctionHouse->RemoveAuction(auction);
+    sAuctionMgr->RemoveAuction(auction);
 
     // Run SQLs
     CharacterDatabase.CommitTransaction(trans);
