+++ src/server/game/World/World.cpp
@@ -31,6 +31,7 @@
 #include "AddonMgr.h"
 #include "ArenaTeamMgr.h"
 #include "AuctionHouseBot.h"
+#include "AuctionHouseCommon.h"
 #include "AuctionHouseMgr.h"
 #include "BattlefieldMgr.h"
 #include "BattlegroundMgr.h"
